--- .eslintignore
@@ -1,8 +1,0 @@
-gulpfile.js
-.eslintrc.js
-.prettierrc.js
-dist
-jsconfig.json
-/dist
-/.pnp.js
-/.yarn/
