+++ .gitattributes
@@ -1,0 +1,7 @@
+.github export-ignore
+FUNDING.yml export-ignore
+
+.gitattributes export-ignore
+README.md export-ignore
+preview.jpg export-ignore
+patchnotes.md export-ignore
