+++ .github/FUNDING.yml
@@ -1,2 +1,2 @@
-ko_fi: mclemente
-custom: https://www.paypal.com/donate/?hosted_button_id=UVFV99HDGE5F8
+github: p4535992
+custom: ["https://www.patreon.com/p4535992"]
