+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,44 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: "[BUG]"
+labels: bug
+assignees: ''
+
+---
+
+**Module Version:** v0.0.0
+
+**Before open any issue**
+
+1) Enable the module setting _"Enable debugging"_
+2) Click F12 go to the _console_ tab 
+3) make the test you want and replicate the error
+4) Go to the tab console open on point 2) and just right click and click 'Save as' and 'Save' or send a screenshot of the exception on the console.
+5) Attach the text file on the github issue with all the logs related tot he module, or send a screenshot of the messages on the console.
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. Go to '...'
+2. Click on '....'
+3. Scroll down to '....'
+4. See error
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Browser:**
+ - 
+
+**Foundry Version:** 
+
+**Game System:** 
+
+**Additional context**
+Add any other context (like other modules installed) about the problem here.
