+++ .github/ISSUE_TEMPLATE/styling.md
@@ -1,0 +1,10 @@
+---
+name: Styling
+about: Request a change or bug related to styles
+title: ''
+labels: styling
+assignees: ''
+
+---
+
+
