+++ package.json
@@ -1,91 +1,64 @@
 {
-  "name": "advanced-macros",
+	"name": "advanced-macros",
 	"title": "Advanced Macros",
-	"description": "Use handlebars templating, recursive macro calls and call macros with arguments or directly from chat.",
-  "version": "1.19.3",
-  "main": "main.js",
-  "scripts": {
-    "publish": "gulp publish --update",
-    "package": "gulp package",
-    "build": "gulp build --dbg",
-    "build:link": "gulp build --dbg && gulp link",
-    "build:test": "gulp build --dbg && gulp link && npm test",
-    "build:prod": "gulp bundle && npm test",
-    "build:watch": "gulp watch",
-    "clean": "gulp clean && gulp link --clean",
-    "test": "jest",
-    "update": "npm install --save-dev @league-of-foundry-developers/foundry-vtt-types@9",
-    "prettier-format": "prettier --config .prettierrc --write \"./src/**/*.{ts,js,json,scss,css}\""
-  },
-  "keywords": [
-    "foundryvtt",
-    "fvtt",
-    "module"
-  ],
-  "author": "",
-  "license": "SEE LICENSE IN LICENSE",
-  "devDependencies": {
-    "@babel/core": "^7.15.0",
-    "@babel/eslint-parser": "^7.15.4",
-    "@league-of-foundry-developers/foundry-vtt-types": "^9.280.0",
-    "@rollup/plugin-node-resolve": "^13.3.0",
-    "@types/archiver": "^5.1.1",
-    "@types/browserify": "^12.0.37",
-    "@types/color": "^3.0.2",
-    "@types/gulp": "^4.0.9",
-    "@types/gulp-less": "^0.0.32",
-    "@types/jest": "^27.4.0",
-    "@types/jquery": "^3.5.14",
-    "@types/yargs": "^17.0.2",
-    "@typescript-eslint/eslint-plugin": "^4.33.0",
-    "@typescript-eslint/parser": "^4.33.0",
-    "@typhonjs-fvtt/eslint-config-foundry.js": "^0.8.0",
-    "archiver": "^5.3.0",
-    "babelify": "^10.0.0",
-    "browserify": "^17.0.0",
-    "color": "3.2.1",
-    "eslint": "^7.32.0",
-    "eslint-plugin-jsdoc": "^39.3.3",
-    "fancy-log": "^2.0.0",
-    "fs": "^0.0.1-security",
-    "gulp": "^4.0.0",
-    "gulp-bro": "^2.0.0",
-    "gulp-changed": "^4.0.3",
-    "gulp-eslint-new": "^1.5.1",
-    "gulp-eslint7": "^0.3.1",
-    "gulp-git": "^2.10.1",
-    "gulp-if": "^3.0.0",
-    "gulp-less": "^5.0.0",
-    "gulp-rename": "^2.0.0",
-    "gulp-replace": "^1.1.3",
-    "gulp-sass": "^5.1.0",
-    "gulp-sourcemaps": "^3.0.0",
-    "gulp-typescript": "^6.0.0-alpha.1",
-    "gulp-uglify": "^3.0.2",
-    "merge-stream": "^2.0.0",
-    "nedb": "^1.8.0",
-    "path": "^0.12.7",
-    "rollup": "^2.77.0",
-    "jest": "^27.5.1",
-    "json-stringify-pretty-compact": "^3.0.0",
-    "sass": "^1.49.11",
-    "through2": "^4.0.2",
-    "ts-jest": "^27.1.3",
-    "ts-node": "^10.2.1",
-    "typescript": "^4.4.2",
-    "uglifyify": "^5.0.2",
-    "vinyl-buffer": "^1.0.1",
-    "vinyl-source-stream": "^2.0.0",
-    "yargs": "^17.1.1"
-  },
-  "dependencies": {
-    "build": "^0.1.4",
-    "csv-parse": "^4.16.3",
-    "gulp-jest": "^4.0.4",
-    "node-sass": "^7.0.1",
-    "prettier": "^2.7.1",
-    "prettier-format": "^3.1.0",
-    "tsify": "^5.0.4",
-    "typescript": "^4.6.3"
-  }
+	"description": "Call macros with arguments directly from chat and execute macros for specific users.",
+	"version": "2.0.6",
+	"main": "module.js",
+	"license": "SEE LICENSE IN LICENSE",
+	"private": true,
+	"type": "module",
+	"author": "",
+	"contributors": [],
+	"imports": {
+		"#runtime/*": "@typhonjs-fvtt/runtime/*",
+		"#standard/*": "@typhonjs-fvtt/svelte-standard/*"
+	},
+	"dependencies": {
+		"@fortawesome/fontawesome-svg-core": "^6.4.0",
+		"@fortawesome/free-solid-svg-icons": "^6.4.0",
+		"@fortawesome/react-fontawesome": "^0.2.0",
+		"@rollup/plugin-node-resolve": "^15.2.1",
+		"@typhonjs-fvtt/runtime": "^0.1.2",
+		"@typhonjs-fvtt/svelte-standard": "^0.1.0",
+		"moment": "^2.29.4",
+		"svelte": "^4.1.2",
+		"svelte-select": "^5.7.0",
+		"svelte-virtual-scroll-list": "^1.1.0"
+	},
+	"devDependencies": {
+        "@babel/eslint-parser": "^7.22.15",
+        "@rollup/plugin-node-resolve": "^15.2.1",
+		"@typhonjs-config/eslint-config": "^0.6.0",
+		"@typhonjs-fvtt/eslint-config-foundry.js": "^0.8.0",
+		"eslint": "^8.46.0",
+		"eslint-plugin-jsdoc": "^46.5.0",
+		"husky": "^8.0.3",
+		"jquery": "^3.6.4",
+		"jsdoc": "^4.0.2",
+		"lint-staged": "^13.2.1",
+		"prettier": "^2.8.7",
+		"sass": "^1.63.6",
+		"svelte-dnd-action": "^0.9.18",
+		"svelte-preprocess": "^5.0.4",
+		"vite": "^4.4.8",
+		"vite-plugin-clean": "^1.0.0",
+		"vite-plugin-run": "^0.4.1",
+		"vite-plugin-static-copy": "^0.16.0"
+	},
+	"browserslist": [
+		">5%",
+		"not IE 11"
+	],
+	"scripts": {
+		"build": "vite build",
+		"dev": "vite",
+		"eslint": "eslint .",
+		"prepare": "husky install",
+		"prettier-format": "prettier --config .prettierrc --write \"./src/**/*.{js,mjs,json,scss,css}\"",
+		"lint": "eslint --ext .js ./src",
+        "lint:fix": "eslint --ext .js ./src --fix"
+	},
+	"lint-staged": {
+		"*.{js,css}": "prettier --write"
+	}
 }
