+++ src/languages/en.json
@@ -1,19 +1,27 @@
 {
-	"advanced-macros": {
-		"MACROS": {
-			"runForSpecificUser": "Execute for Specific User",
-			"runForSpecificUserTooltip": "This will cause the macro to be executed by the GM for a specific player connected.\nFor security reasons, only applies to GM-created macros with no other owners.",
-			"runForEveryone": "Everyone",
-			"runForEveryoneElse": "Everyone else",
-			"none": "None",
-			"runAsWorldScript": "World Script",
-			"runAsWorldScriptTooltip": "This will cause the macro to be executed automatically be the world once it's loaded.\nFor security reasons, only applies to GM-created macros with no other owners."
-		},
-		"setting": {
-			"legacySlashCommand": {
-				"name": "Chat Slash Command",
-				"hint": "Ignores the need to use the /macro command to call macros on the chat. Warning: this might cause conflict with your system or other modules."
-			}
-		}
-	}
+  "advanced-macros": {
+    "MACROS": {
+      "runForSpecificUser": "Execute for Specific User",
+      "runForSpecificUserTooltip": "This will cause the macro to be executed by the GM for a specific player connected.\nFor security reasons, the \"GM\" option only applies to GM-created macros with no other owners.",
+      "runForEveryoneOption": "👪 Everyone",
+      "runForEveryoneElseOption": "👥 Everyone else",
+      "runAsNoneOption": "None",
+      "runAsWorldScriptOption": "🌍 World Script",
+      "runAsGMOption": "💣 GM",
+      "runAsMySelfOption": "👤 My Self ({name})",
+      "runAsWorldScript": "Execute macro as World Script ?",
+      "runAsWorldScriptTooltip": "This will cause the macro to be executed automatically be the world once it's loaded.\nNOTE: This setting wins over the \"specific user\" value of the dropdown if set",
+      "runAsGM": "Execute macro as GM ?",
+      "runAsGMTooltip": "This will cause the macro to be executed by the GM when players execute it.\nFor security reasons, only applies to GM-created macros with no other owners.\nNOTE: This setting wins over the \"specific user\" value of the dropdown if set",
+      "runAsPersonalTitle": "Run as a personal macro",
+      "runAsPersonalHint": "",
+      "runAsPersonal": "Execute macro as Personal Script ?"
+    },
+    "setting": {
+      "legacySlashCommand": {
+        "name": "Chat Slash Command",
+        "hint": "Ignores the need to use the /macro command to call macros on the chat. Warning: this might cause conflict with your system or other modules."
+      }
+    }
+  }
 }
